perpetua/pyproject.toml
[project]
name = "perpetua"
version = "0.1.2"
description = "Perpetua is a python package that creates agents that are context aware of your codebase."
authors = [
    {name = "Sami Khayatei Houssaini",email = "sami_houssaini@berkeley.edu"}
]
readme = "README.md"
requires-python = ">=3.10,<=3.14"
dependencies = [
    "typer (>=0.20.0,<0.21.0)",
    "langgraph (>=1.0.4,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "langchain (>=1.1.0,<2.0.0)",
    "langchain-google-genai (>=3.2.0,<4.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain-core (>=1.1.0,<2.0.0)",
    "langchain-milvus (>=0.3.0,<0.4.0)",
    "rich (>=14.2.0,<15.0.0)",
    "langchain-tavily (>=0.2.13,<0.3.0)",
    "langchain-ollama (>=0.1.0)",
    "langchain-community (>=0.0.20,<1.0.0)",
    "tree-sitter (>=0.21.0,<0.22.0)",
    "tree-sitter-languages (>=1.9.0,<1.10.0)",
    "langmem (>=0.0.30,<1.0.0)",
    "networkx (>=3.0.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "perpetua", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
perpetua = "perpetua.app:app"